refactor handle visibility changes into helper module
