+++ packages/helpers/src/index.ts
@@ -19,6 +19,7 @@ export { default as forceColors} from './force-colors'
 export { default as forceIcons} from './force-icons'
 export { default as getSelectedItem } from './getSelectedItem'
 export { default as handleChannelUpdates, handleAllInitialChannelMessages, clipInfoMEssage } from './handle-channel-updates'
+export { default as handleVisibilityChanges } from './handle-visibility-changes'
 export { default as hexNamed } from './hex-named'
 export { default as isChatChannel } from './is-chat-channel'
 export { default as isArrayOfObject } from './is-array-of-object'
